+++ updatehub/states.go
@@ -9,14 +9,7 @@
 package updatehub
 
 import (
-	"errors"
-	"fmt"
-	"sync"
-	"time"
-
-	"github.com/OSSystems/pkg/log"
 	"github.com/UpdateHub/updatehub/metadata"
-	"github.com/spf13/afero"
 )
 
 // UpdateHubState holds the possible states for the agent
@@ -122,532 +115,7 @@ func StateToString(status UpdateHubState) string {
 	return statusNames[status]
 }
 
-// ErrorState is the State interface implementation for the UpdateHubStateError
-type ErrorState struct {
-	BaseState
-	cause UpdateHubErrorReporter
-	ReportableState
-
-	updateMetadata *metadata.UpdateMetadata
-}
-
-// UpdateMetadata is the ReportableState interface implementation
-func (state *ErrorState) UpdateMetadata() *metadata.UpdateMetadata {
-	return state.updateMetadata
-}
-
-// Handle for ErrorState calls "panic" if the error is fatal or
-// triggers a poll state otherwise
-func (state *ErrorState) Handle(uh *UpdateHub) (State, bool) {
-	log.Warn(state.cause)
-
-	if state.cause.IsFatal() {
-		return NewExitState(1), false
-	}
-
-	return NewIdleState(), false
-}
-
-// ToMap is for the State interface implementation
-func (state *ErrorState) ToMap() map[string]interface{} {
-	m := state.BaseState.ToMap()
-	m["error"] = state.cause.Error()
-	return m
-}
-
-// NewErrorState creates a new ErrorState from a UpdateHubErrorReporter
-func NewErrorState(updateMetadata *metadata.UpdateMetadata, err UpdateHubErrorReporter) State {
-	if err == nil {
-		err = NewFatalError(errors.New("generic error"))
-	}
-
-	return &ErrorState{
-		BaseState:      BaseState{id: UpdateHubStateError},
-		cause:          err,
-		updateMetadata: updateMetadata,
-	}
-}
-
 // ReportableState interface describes the necessary operations for a State to be reportable
 type ReportableState interface {
 	UpdateMetadata() *metadata.UpdateMetadata
-}
-
-// IdleState is the State interface implementation for the UpdateHubStateIdle
-type IdleState struct {
-	BaseState
-	CancellableState
-}
-
-// ID returns the state id
-func (state *IdleState) ID() UpdateHubState {
-	return state.id
-}
-
-// Cancel cancels a state if it is cancellable
-func (state *IdleState) Cancel(ok bool, nextState State) bool {
-	return state.CancellableState.Cancel(ok, nextState)
-}
-
-// Handle for IdleState
-func (state *IdleState) Handle(uh *UpdateHub) (State, bool) {
-	if !uh.Settings.PollingEnabled {
-		state.Wait()
-		return state.NextState(), false
-	}
-
-	now := time.Now()
-
-	if uh.Settings.ExtraPollingInterval > 0 {
-		extraPollTime := uh.Settings.LastPoll.Add(uh.Settings.ExtraPollingInterval)
-
-		log.Info("ExtraPoll received: ", extraPollTime)
-
-		if extraPollTime.Before(now) {
-			return NewUpdateProbeState(), false
-		}
-	}
-
-	return NewPollState(uh.Settings.PollingInterval), false
-}
-
-// NewIdleState creates a new IdleState
-func NewIdleState() *IdleState {
-	state := &IdleState{
-		BaseState:        BaseState{id: UpdateHubStateIdle},
-		CancellableState: CancellableState{cancel: make(chan bool)},
-	}
-
-	return state
-}
-
-// PollState is the State interface implementation for the UpdateHubStatePoll
-type PollState struct {
-	BaseState
-	CancellableState
-
-	interval   time.Duration
-	ticksCount int64
-}
-
-// ID returns the state id
-func (state *PollState) ID() UpdateHubState {
-	return state.id
-}
-
-// Cancel cancels a state if it is cancellable
-func (state *PollState) Cancel(ok bool, nextState State) bool {
-	return state.CancellableState.Cancel(ok, nextState)
-}
-
-// Handle for PollState encapsulates the polling logic
-func (state *PollState) Handle(uh *UpdateHub) (State, bool) {
-	var nextState State
-
-	nextState = state
-
-	if state.interval < uh.TimeStep {
-		finalErr := fmt.Errorf("polling interval (%s) must be greater than '%s', using %s instead", state.interval, uh.TimeStep, uh.TimeStep)
-		log.Warn(finalErr)
-		state.interval = uh.TimeStep
-	}
-
-	go func() {
-		ticks := state.ticksCount
-
-	polling:
-		for {
-			ticker := time.NewTicker(uh.TimeStep)
-
-			defer ticker.Stop()
-
-			select {
-			case <-ticker.C:
-				ticks++
-
-				if ticks > 0 && ticks%int64(state.interval.Seconds()/uh.TimeStep.Seconds()) == 0 {
-					nextState = NewUpdateProbeState()
-					break polling
-				}
-			case <-state.cancel:
-				break
-			}
-		}
-
-		state.Cancel(true, nextState)
-
-		state.ticksCount = ticks
-	}()
-
-	state.Wait()
-
-	// state cancelled
-	if state.NextState() != nil {
-		return state.NextState(), true
-	}
-
-	return nextState, false
-}
-
-// NewPollState creates a new PollState
-func NewPollState(pollingInterval time.Duration) *PollState {
-	state := &PollState{
-		BaseState:        BaseState{id: UpdateHubStatePoll},
-		CancellableState: CancellableState{cancel: make(chan bool, 1)},
-	}
-
-	state.interval = pollingInterval
-
-	return state
-}
-
-// UpdateProbeState is the State interface implementation for the UpdateHubStateUpdateProbe
-type UpdateProbeState struct {
-	BaseState
-}
-
-// ID returns the state id
-func (state *UpdateProbeState) ID() UpdateHubState {
-	return state.id
-}
-
-// Handle for UpdateProbeState executes a ProbeUpdate procedure and
-// proceed to download the update if there is one. It goes back to the
-// polling state otherwise.
-func (state *UpdateProbeState) Handle(uh *UpdateHub) (State, bool) {
-	updateMetadata, extraPoll := uh.Controller.ProbeUpdate(uh.Settings.PollingRetries)
-
-	// Reset polling retries in case of ProbeUpdate success
-	if extraPoll != -1 {
-		uh.Settings.PollingRetries = 0
-	}
-
-	uh.Settings.LastPoll = time.Now()
-	uh.Settings.ExtraPollingInterval = 0
-
-	if updateMetadata != nil {
-		packageUID := updateMetadata.PackageUID()
-		if packageUID == uh.lastInstalledPackageUID {
-			return NewWaitingForRebootState(updateMetadata), false
-		}
-
-		return NewDownloadingState(updateMetadata, &ProgressTrackerImpl{}), false
-	}
-
-	if extraPoll > 0 {
-		now := time.Now()
-		nextPoll := time.Unix(uh.Settings.FirstPoll.Unix(), 0)
-		extraPollTime := now.Add(extraPoll)
-
-		for nextPoll.Before(now) {
-			nextPoll = nextPoll.Add(uh.Settings.PollingInterval)
-		}
-
-		if extraPollTime.Before(nextPoll) {
-			uh.Settings.ExtraPollingInterval = extraPoll
-
-			poll := NewPollState(uh.Settings.PollingInterval)
-			poll.interval = extraPoll
-
-			return poll, false
-		}
-	}
-
-	// Increment the number of polling retries in case of ProbeUpdate failure
-	uh.Settings.PollingRetries++
-
-	return NewIdleState(), false
-}
-
-// NewUpdateProbeState creates a new UpdateProbeState
-func NewUpdateProbeState() *UpdateProbeState {
-	state := &UpdateProbeState{
-		BaseState: BaseState{id: UpdateHubStateUpdateProbe},
-	}
-
-	return state
-}
-
-// DownloadingState is the State interface implementation for the UpdateHubStateDownloading
-type DownloadingState struct {
-	BaseState
-	CancellableState
-	ReportableState
-	ProgressTracker
-
-	updateMetadata *metadata.UpdateMetadata
-}
-
-// ID returns the state id
-func (state *DownloadingState) ID() UpdateHubState {
-	return state.id
-}
-
-// Cancel cancels a state if it is cancellable
-func (state *DownloadingState) Cancel(ok bool, nextState State) bool {
-	state.CancellableState.Cancel(ok, nextState)
-	return ok
-}
-
-// UpdateMetadata is the ReportableState interface implementation
-func (state *DownloadingState) UpdateMetadata() *metadata.UpdateMetadata {
-	return state.updateMetadata
-}
-
-// Handle for DownloadingState starts the objects downloads. It goes
-// to the installing state if successfull. It goes back to the error
-// state otherwise.
-func (state *DownloadingState) Handle(uh *UpdateHub) (State, bool) {
-	var err error
-	var nextState State
-
-	nextState = state
-
-	progressChan := make(chan int, 10)
-
-	m := sync.Mutex{}
-	m.Lock()
-
-	go func() {
-		m.Lock()
-		defer m.Unlock()
-
-		err = uh.Controller.DownloadUpdate(state.updateMetadata, state.cancel, progressChan)
-		close(progressChan)
-	}()
-
-	m.Unlock()
-	for p := range progressChan {
-		state.ProgressTracker.SetProgress(p)
-	}
-
-	if err != nil {
-		nextState = NewErrorState(state.updateMetadata, NewTransientError(err))
-	} else {
-		nextState = NewDownloadedState(state.updateMetadata)
-	}
-
-	// state cancelled
-	if state.NextState() != nil {
-		return state.NextState(), true
-	}
-
-	return nextState, false
-}
-
-// ToMap is for the State interface implementation
-func (state *DownloadingState) ToMap() map[string]interface{} {
-	m := state.BaseState.ToMap()
-	m["progress"] = state.ProgressTracker.GetProgress()
-	return m
-}
-
-// NewDownloadingState creates a new DownloadingState from a metadata.UpdateMetadata
-func NewDownloadingState(updateMetadata *metadata.UpdateMetadata, pti ProgressTracker) *DownloadingState {
-	state := &DownloadingState{
-		BaseState:       BaseState{id: UpdateHubStateDownloading},
-		updateMetadata:  updateMetadata,
-		ProgressTracker: pti,
-	}
-
-	return state
-}
-
-// DownloadedState is the State interface implementation for the UpdateHubStateDownloaded
-type DownloadedState struct {
-	BaseState
-	ReportableState
-
-	updateMetadata *metadata.UpdateMetadata
-}
-
-// ID returns the state id
-func (state *DownloadedState) ID() UpdateHubState {
-	return state.id
-}
-
-// UpdateMetadata is the ReportableState interface implementation
-func (state *DownloadedState) UpdateMetadata() *metadata.UpdateMetadata {
-	return state.updateMetadata
-}
-
-// Handle for DownloadedState just returns a new installing state
-func (state *DownloadedState) Handle(uh *UpdateHub) (State, bool) {
-	return NewInstallingState(state.updateMetadata, &ProgressTrackerImpl{}, uh.Store), false
-}
-
-// NewDownloadedState creates a new DownloadedState
-func NewDownloadedState(updateMetadata *metadata.UpdateMetadata) *DownloadedState {
-	state := &DownloadedState{
-		BaseState:      BaseState{id: UpdateHubStateDownloaded},
-		updateMetadata: updateMetadata,
-	}
-
-	return state
-}
-
-// InstallingState is the State interface implementation for the UpdateHubStateInstalling
-type InstallingState struct {
-	BaseState
-	ReportableState
-	ProgressTracker
-	FileSystemBackend afero.Fs
-	updateMetadata    *metadata.UpdateMetadata
-}
-
-// ID returns the state id
-func (state *InstallingState) ID() UpdateHubState {
-	return state.id
-}
-
-// Handle for InstallingState implements the installation process itself
-func (state *InstallingState) Handle(uh *UpdateHub) (State, bool) {
-	packageUID := state.updateMetadata.PackageUID()
-	if packageUID == uh.lastInstalledPackageUID {
-		return NewWaitingForRebootState(state.updateMetadata), false
-	}
-
-	// register the packageUID at the start so it won't redo the
-	// operations in case of an install error occurs
-	uh.lastInstalledPackageUID = packageUID
-
-	var err error
-
-	progressChan := make(chan int, 10)
-
-	m := sync.Mutex{}
-	m.Lock()
-
-	go func() {
-		m.Lock()
-		defer m.Unlock()
-
-		err = uh.Controller.InstallUpdate(state.updateMetadata, progressChan)
-		close(progressChan)
-	}()
-
-	m.Unlock()
-	for p := range progressChan {
-		state.ProgressTracker.SetProgress(p)
-	}
-
-	if err != nil {
-		return NewErrorState(state.updateMetadata, NewTransientError(err)), false
-	}
-
-	return NewInstalledState(state.updateMetadata), false
-}
-
-// ToMap is for the State interface implementation
-func (state *InstallingState) ToMap() map[string]interface{} {
-	m := state.BaseState.ToMap()
-	m["progress"] = state.ProgressTracker.GetProgress()
-	return m
-}
-
-// UpdateMetadata is the ReportableState interface implementation
-func (state *InstallingState) UpdateMetadata() *metadata.UpdateMetadata {
-	return state.updateMetadata
-}
-
-// NewInstallingState creates a new InstallingState
-func NewInstallingState(
-	updateMetadata *metadata.UpdateMetadata,
-	pti ProgressTracker,
-	fsb afero.Fs) *InstallingState {
-	state := &InstallingState{
-		BaseState:         BaseState{id: UpdateHubStateInstalling},
-		updateMetadata:    updateMetadata,
-		FileSystemBackend: fsb,
-		ProgressTracker:   pti,
-	}
-
-	return state
-}
-
-// WaitingForRebootState is the State interface implementation for the UpdateHubStateWaitingForReboot
-type WaitingForRebootState struct {
-	BaseState
-	ReportableState
-
-	updateMetadata *metadata.UpdateMetadata
-}
-
-// ID returns the state id
-func (state *WaitingForRebootState) ID() UpdateHubState {
-	return state.id
-}
-
-// Handle for WaitingForRebootState tells us that an installation has
-// been made and it is waiting for a reboot
-func (state *WaitingForRebootState) Handle(uh *UpdateHub) (State, bool) {
-	return NewIdleState(), false
-}
-
-// UpdateMetadata is the ReportableState interface implementation
-func (state *WaitingForRebootState) UpdateMetadata() *metadata.UpdateMetadata {
-	return state.updateMetadata
-}
-
-// NewWaitingForRebootState creates a new WaitingForRebootState
-func NewWaitingForRebootState(updateMetadata *metadata.UpdateMetadata) *WaitingForRebootState {
-	state := &WaitingForRebootState{
-		BaseState:      BaseState{id: UpdateHubStateWaitingForReboot},
-		updateMetadata: updateMetadata,
-	}
-
-	return state
-}
-
-// InstalledState is the State interface implementation for the UpdateHubStateInstalled
-type InstalledState struct {
-	BaseState
-	ReportableState
-
-	updateMetadata *metadata.UpdateMetadata
-}
-
-// ID returns the state id
-func (state *InstalledState) ID() UpdateHubState {
-	return state.id
-}
-
-// Handle for InstalledState implements the installation process itself
-func (state *InstalledState) Handle(uh *UpdateHub) (State, bool) {
-	return NewIdleState(), false
-}
-
-// UpdateMetadata is the ReportableState interface implementation
-func (state *InstalledState) UpdateMetadata() *metadata.UpdateMetadata {
-	return state.updateMetadata
-}
-
-// NewInstalledState creates a new InstalledState
-func NewInstalledState(updateMetadata *metadata.UpdateMetadata) *InstalledState {
-	state := &InstalledState{
-		BaseState:      BaseState{id: UpdateHubStateInstalled},
-		updateMetadata: updateMetadata,
-	}
-
-	return state
-}
-
-// ExitState is the final state of the state machine
-type ExitState struct {
-	BaseState
-
-	exitCode int
-}
-
-// NewExitState creates a new ExitState
-func NewExitState(exitCode int) *ExitState {
-	return &ExitState{
-		BaseState: BaseState{id: UpdateHubStateExit},
-		exitCode:  exitCode,
-	}
-}
-
-// Handle for ExitState
-func (state *ExitState) Handle(uh *UpdateHub) (State, bool) {
-	panic("ExitState handler should not be called")
 }
